requirements.txt
mysql-connector
discord.py
buttons
requests
